typechecker: refactor + dispatch
